% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competition.R
\docType{data}
\name{Challenge}
\alias{Challenge}
\title{Machine Learning Challenge.}
\format{\code{\link{R6Class}} object.}
\usage{
Challenge
}
\description{
Create a Machine Learning challenge objects that should be solved in a given time limit.
Currently only classification tasks with accuracy as performance are supported.
}
\section{Usage}{

\preformatted{
chall = Challenge$new(id = 3, difficulty = "easy", time.limit = 3600)
lrn = makeLearner("classif.rpart")
chall$submit(lrn)
}
}

\section{Arguments}{

\strong{For Challenge$new()}:
\describe{
\item{\code{id}}{[\code{integer(1)} | \code{NULL} | \code{"daily"}]\cr
  OpenML task.id used as challenge task. If \code{NULL} a random task from the
  \code{OpenML-CC18} benchmarking suit is sampled.
  A daily challenge can be selected by setting \code{id} to \code{"daily"}.
}
\item{\code{difficulty}}{[\code{character(1) | numeric(1)}]\cr
  Define the difficulty of the Challenge. \code{very easy} is the 0.5 quantile of runs submitted to OpenML,
  \code{easy} the 0.6, \code{medium} the 0.8, \code{hard} the 0.9 and \code{very hard} the 0.99 quantile.
  Alternatively the quantile can be directly specified as a numer value between 0 and 1.
  Default is \code{easy}.
}
\item{\code{time.limit}}{[\code{numeric(1)}]\cr
  Time limit, starting from the creation of the Challenge object until it is solved.
  Default is \code{1800}, i.e., half an hour.
}
}
\strong{For Challenge$submit()}:
\describe{
\item{\code{lrn}}{[\code{\link[mlr]{Learner}}]\cr
  Learner submitted to solved the challenge. Preprocessing on the task can be done with \pkg{mlrCPO}.
}
}
}

\section{Fields}{

\describe{
\item{\code{id} [\code{integer(1)}]}{
  id of \pkg{OpenML} task used in challenge.
}
\item{\code{difficulty} [\code{numeric(1)}]}{
  Difficulty of the challenge.
}
\item{\code{goal} [\code{numeric(1)}]}{
  Threshold value to beat with \code{Challenge$submit(lrn)}.
}
\item{\code{task} [\code{\link[mlr]{Task}}]}{
  The Data for the challenge, useful for exploratory analysis.
}
\item{\code{time.limit}}{[\code{numeric(1)}]\cr
  See above.
}
}
}

\examples{
chall = Challenge$new(id = 3, difficulty = "easy", time.limit = 3600)
lrn = makeLearner("classif.rpart")
chall$submit(lrn)
}
\keyword{datasets}
